+++ src/validator.js
@@ -8,6 +8,8 @@ const {
   utf8ToBuffer,
   encodeTag,
   bytes32Regex,
+  parseRLC,
+  parseEth,
 } = require('./utils');
 const { paramsKeyName, storageProviders } = require('./params-utils');
 const { teePostComputeDefaults } = require('./secrets-utils');
@@ -41,6 +43,62 @@ const hexnumberSchema = () => string()
   );
 
 const uint256Schema = () => stringNumberSchema({ message: '${originalValue} is not a valid uint256' });
+
+const amontErrorMessage = ({ originalValue }) => `${
+  Array.isArray(originalValue) ? originalValue.join(' ') : originalValue
+} is not a valid amount`;
+
+const nRlcAmountSchema = ({ defaultUnit = 'nRLC' } = {}) => string()
+  .transform((value, originalValue) => {
+    if (Array.isArray(originalValue)) {
+      if (originalValue.length > 2) {
+        throw new ValidationError(amontErrorMessage({ originalValue }));
+      }
+      if (originalValue.length === 2 && originalValue[1]) {
+        return `${originalValue[0]} ${originalValue[1]}`;
+      }
+      return `${originalValue[0]}`;
+    }
+    return value;
+  })
+  .transform((value) => {
+    const trimed = value.replace(/^0+/, '');
+    return trimed.length > 0 ? trimed : '0';
+  })
+  .transform((value, originalValue) => {
+    try {
+      return parseRLC(value, defaultUnit).toString();
+    } catch (e) {
+      throw new ValidationError(amontErrorMessage({ originalValue }));
+    }
+  })
+  .matches(/^[0-9]*$/, amontErrorMessage);
+
+const weiAmountSchema = ({ defaultUnit = 'wei' } = {}) => string()
+  .transform((value, originalValue) => {
+    if (Array.isArray(originalValue)) {
+      if (originalValue.length > 2) {
+        throw new ValidationError(amontErrorMessage({ originalValue }));
+      }
+      if (originalValue.length === 2 && originalValue[1]) {
+        return `${originalValue[0]} ${originalValue[1]}`;
+      }
+      return `${originalValue[0]}`;
+    }
+    return value;
+  })
+  .transform((value) => {
+    const trimed = value.replace(/^0+/, '');
+    return trimed.length > 0 ? trimed : '0';
+  })
+  .transform((value, originalValue) => {
+    try {
+      return parseEth(value, defaultUnit).toString();
+    } catch (e) {
+      throw new ValidationError(amontErrorMessage({ originalValue }));
+    }
+  })
+  .matches(/^[0-9]*$/, amontErrorMessage);
 
 const chainIdSchema = () => stringNumberSchema({ message: '${originalValue} is not a valid chainId' });
 
@@ -272,14 +330,14 @@ const tagSchema = () => mixed()
 const apporderSchema = opt => object(
   {
     app: addressSchema(opt).required(),
-    appprice: uint256Schema().required(),
+    appprice: nRlcAmountSchema().required(),
     volume: uint256Schema().required(),
     tag: tagSchema().required(),
     datasetrestrict: addressSchema(opt).required(),
     workerpoolrestrict: addressSchema(opt).required(),
     requesterrestrict: addressSchema(opt).required(),
   },
-  '${originalValue} is not a valid signed apporder',
+  '${originalValue} is not a valid apporder',
 );
 
 const saltedApporderSchema = opt => apporderSchema(opt).shape(
@@ -295,14 +353,14 @@ const signedApporderSchema = opt => saltedApporderSchema(opt).shape(
 const datasetorderSchema = opt => object(
   {
     dataset: addressSchema(opt).required(),
-    datasetprice: uint256Schema().required(),
+    datasetprice: nRlcAmountSchema().required(),
     volume: uint256Schema().required(),
     tag: tagSchema().required(),
     apprestrict: addressSchema(opt).required(),
     workerpoolrestrict: addressSchema(opt).required(),
     requesterrestrict: addressSchema(opt).required(),
   },
-  '${originalValue} is not a valid signed datasetorder',
+  '${originalValue} is not a valid datasetorder',
 );
 
 const saltedDatasetorderSchema = opt => datasetorderSchema(opt).shape(
@@ -318,7 +376,7 @@ const signedDatasetorderSchema = opt => saltedDatasetorderSchema(opt).shape(
 const workerpoolorderSchema = opt => object(
   {
     workerpool: addressSchema(opt).required(),
-    workerpoolprice: uint256Schema().required(),
+    workerpoolprice: nRlcAmountSchema().required(),
     volume: uint256Schema().required(),
     tag: tagSchema().required(),
     category: catidSchema().required(),
@@ -327,7 +385,7 @@ const workerpoolorderSchema = opt => object(
     datasetrestrict: addressSchema(opt).required(),
     requesterrestrict: addressSchema(opt).required(),
   },
-  '${originalValue} is not a valid signed workerpoolorder',
+  '${originalValue} is not a valid workerpoolorder',
 );
 
 const saltedWorkerpoolorderSchema = opt => workerpoolorderSchema(opt).shape(
@@ -337,17 +395,17 @@ const saltedWorkerpoolorderSchema = opt => workerpoolorderSchema(opt).shape(
 
 const signedWorkerpoolorderSchema = opt => saltedWorkerpoolorderSchema(opt).shape(
   signed(),
-  '${originalValue} is not a valid salted workerpoolorder',
+  '${originalValue} is not a valid signed workerpoolorder',
 );
 
 const requestorderSchema = opt => object(
   {
     app: addressSchema(opt).required(),
-    appmaxprice: uint256Schema().required(),
+    appmaxprice: nRlcAmountSchema().required(),
     dataset: addressSchema(opt).required(),
-    datasetmaxprice: uint256Schema().required(),
+    datasetmaxprice: nRlcAmountSchema().required(),
     workerpool: addressSchema(opt).required(),
-    workerpoolmaxprice: uint256Schema().required(),
+    workerpoolmaxprice: nRlcAmountSchema().required(),
     requester: addressSchema(opt).required(),
     volume: uint256Schema().required(),
     tag: tagSchema().required(),
@@ -357,7 +415,7 @@ const requestorderSchema = opt => object(
     callback: addressSchema(opt).required(),
     params: paramsSchema(),
   },
-  '${originalValue} is not a valid signed requestorder',
+  '${originalValue} is not a valid requestorder',
 );
 
 const saltedRequestorderSchema = opt => requestorderSchema(opt).shape(
@@ -420,6 +478,8 @@ module.exports = {
   throwIfMissing,
   stringSchema: string,
   uint256Schema,
+  nRlcAmountSchema,
+  weiAmountSchema,
   addressSchema,
   bytes32Schema,
   apporderSchema,