+++ test/lib.test.js
@@ -1354,6 +1354,41 @@ describe('[wallet]', () => {
       true,
     );
   });
+  test('wallet.sendETH() (specified unit)', async () => {
+    const signer = utils.getSignerFromPrivateKey(tokenChainUrl, PRIVATE_KEY);
+    const iexec = new IExec(
+      {
+        ethProvider: signer,
+        chainId: networkId,
+      },
+      {
+        hubAddress,
+        isNative: false,
+      },
+    );
+    const initialBalance = await iexec.wallet.checkBalances(ADDRESS);
+    const receiverInitialBalance = await iexec.wallet.checkBalances(
+      POOR_ADDRESS3,
+    );
+    const txHash = await iexec.wallet.sendETH('0.5 gwei', POOR_ADDRESS3);
+    const finalBalance = await iexec.wallet.checkBalances(ADDRESS);
+    const receiverFinalBalance = await iexec.wallet.checkBalances(
+      POOR_ADDRESS3,
+    );
+    expect(txHash).toMatch(bytes32Regex);
+    expect(
+      finalBalance.wei.add(new BN('500000000')).lt(initialBalance.wei),
+    ).toBe(true);
+    expect(finalBalance.nRLC.eq(initialBalance.nRLC)).toBe(true);
+    expect(
+      receiverFinalBalance.wei.eq(
+        receiverInitialBalance.wei.add(new BN('500000000')),
+      ),
+    ).toBe(true);
+    expect(receiverFinalBalance.nRLC.eq(receiverInitialBalance.nRLC)).toBe(
+      true,
+    );
+  });
   test('wallet.sendETH() (throw on native)', async () => {
     const signer = utils.getSignerFromPrivateKey(nativeChainUrl, PRIVATE_KEY);
     const iexec = new IExec(
@@ -1395,6 +1430,42 @@ describe('[wallet]', () => {
     expect(finalBalance.wei.lt(initialBalance.wei)).toBe(true);
     expect(finalBalance.nRLC.add(new BN(5)).eq(initialBalance.nRLC)).toBe(true);
     expect(receiverFinalBalance.wei.eq(receiverInitialBalance.wei)).toBe(true);
+    expect(
+      receiverFinalBalance.nRLC.eq(receiverInitialBalance.nRLC.add(new BN(5))),
+    ).toBe(true);
+  });
+  test('wallet.sendRLC() (specified unit)', async () => {
+    const signer = utils.getSignerFromPrivateKey(tokenChainUrl, PRIVATE_KEY);
+    const iexec = new IExec(
+      {
+        ethProvider: signer,
+        chainId: networkId,
+      },
+      {
+        hubAddress,
+        isNative: false,
+      },
+    );
+    const initialBalance = await iexec.wallet.checkBalances(ADDRESS);
+    const receiverInitialBalance = await iexec.wallet.checkBalances(
+      POOR_ADDRESS3,
+    );
+    const txHash = await iexec.wallet.sendRLC('0.5 RLC', POOR_ADDRESS3);
+    const finalBalance = await iexec.wallet.checkBalances(ADDRESS);
+    const receiverFinalBalance = await iexec.wallet.checkBalances(
+      POOR_ADDRESS3,
+    );
+    expect(txHash).toMatch(bytes32Regex);
+    expect(finalBalance.wei.lt(initialBalance.wei)).toBe(true);
+    expect(
+      finalBalance.nRLC.add(new BN('500000000')).eq(initialBalance.nRLC),
+    ).toBe(true);
+    expect(receiverFinalBalance.wei.eq(receiverInitialBalance.wei)).toBe(true);
+    expect(
+      receiverFinalBalance.nRLC.eq(
+        receiverInitialBalance.nRLC.add(new BN('500000000')),
+      ),
+    ).toBe(true);
   });
   test('wallet.sendRLC() (native)', async () => {
     const signer = utils.getSignerFromPrivateKey(nativeChainUrl, PRIVATE_KEY);
@@ -1430,6 +1501,47 @@ describe('[wallet]', () => {
     expect(
       receiverFinalBalance.wei
         .sub(new BN(5).mul(new BN(1000000000)))
+        .eq(receiverInitialBalance.wei),
+    ).toBe(true);
+  });
+  test('wallet.sendRLC() (native, specified unit)', async () => {
+    const signer = utils.getSignerFromPrivateKey(nativeChainUrl, PRIVATE_KEY);
+    const iexec = new IExec(
+      {
+        ethProvider: signer,
+        chainId: networkId,
+      },
+      {
+        hubAddress: nativeHubAddress,
+        isNative: true,
+      },
+    );
+    const initialBalance = await iexec.wallet.checkBalances(ADDRESS);
+    const receiverInitialBalance = await iexec.wallet.checkBalances(
+      POOR_ADDRESS3,
+    );
+    const txHash = await iexec.wallet.sendRLC('0.000005 RLC', POOR_ADDRESS3);
+    const finalBalance = await iexec.wallet.checkBalances(ADDRESS);
+    const receiverFinalBalance = await iexec.wallet.checkBalances(
+      POOR_ADDRESS3,
+    );
+    expect(txHash).toMatch(bytes32Regex);
+    expect(finalBalance.nRLC.add(new BN(5000)).eq(initialBalance.nRLC)).toBe(
+      true,
+    );
+    expect(
+      finalBalance.wei
+        .add(new BN(5000).mul(new BN(1000000000)))
+        .eq(initialBalance.wei),
+    ).toBe(true);
+    expect(
+      receiverFinalBalance.nRLC
+        .sub(new BN(5000))
+        .eq(receiverInitialBalance.nRLC),
+    ).toBe(true);
+    expect(
+      receiverFinalBalance.wei
+        .sub(new BN(5000).mul(new BN(1000000000)))
         .eq(receiverInitialBalance.wei),
     ).toBe(true);
   });
@@ -1691,6 +1803,39 @@ describe('[account]', () => {
       true,
     );
   });
+  test('account.deposit() (token, specified unit)', async () => {
+    const signer = utils.getSignerFromPrivateKey(tokenChainUrl, PRIVATE_KEY);
+    const iexec = new IExec(
+      {
+        ethProvider: signer,
+        chainId: networkId,
+      },
+      {
+        hubAddress,
+        isNative: false,
+      },
+    );
+    const accountInitialBalance = await iexec.account.checkBalance(ADDRESS);
+    const walletInitialBalance = await iexec.wallet.checkBalances(ADDRESS);
+    const res = await iexec.account.deposit('0.005 RLC');
+    const accountFinalBalance = await iexec.account.checkBalance(ADDRESS);
+    const walletFinalBalance = await iexec.wallet.checkBalances(ADDRESS);
+    expect(res.txHash).toMatch(bytes32Regex);
+    expect(res.amount).toBe('5000000');
+    expect(
+      accountFinalBalance.stake
+        .sub(new BN('5000000'))
+        .eq(accountInitialBalance.stake),
+    ).toBe(true);
+    expect(
+      walletFinalBalance.nRLC
+        .add(new BN('5000000'))
+        .eq(walletInitialBalance.nRLC),
+    ).toBe(true);
+    expect(accountFinalBalance.locked.eq(accountInitialBalance.locked)).toBe(
+      true,
+    );
+  });
   test('account.deposit() (token, exceed wallet balance)', async () => {
     const signer = utils.getSignerFromPrivateKey(tokenChainUrl, PRIVATE_KEY);
     const iexec = new IExec(
@@ -1743,6 +1888,39 @@ describe('[account]', () => {
     ).toBe(true);
     expect(
       walletFinalBalance.nRLC.add(new BN(5)).eq(walletInitialBalance.nRLC),
+    ).toBe(true);
+    expect(accountFinalBalance.locked.eq(accountInitialBalance.locked)).toBe(
+      true,
+    );
+  });
+  test('account.deposit() (native, specified unit)', async () => {
+    const signer = utils.getSignerFromPrivateKey(nativeChainUrl, PRIVATE_KEY);
+    const iexec = new IExec(
+      {
+        ethProvider: signer,
+        chainId: networkId,
+      },
+      {
+        hubAddress: nativeHubAddress,
+        isNative: true,
+      },
+    );
+    const accountInitialBalance = await iexec.account.checkBalance(ADDRESS);
+    const walletInitialBalance = await iexec.wallet.checkBalances(ADDRESS);
+    const res = await iexec.account.deposit('0.005 RLC');
+    const accountFinalBalance = await iexec.account.checkBalance(ADDRESS);
+    const walletFinalBalance = await iexec.wallet.checkBalances(ADDRESS);
+    expect(res.txHash).toMatch(bytes32Regex);
+    expect(res.amount).toBe('5000000');
+    expect(
+      accountFinalBalance.stake
+        .sub(new BN('5000000'))
+        .eq(accountInitialBalance.stake),
+    ).toBe(true);
+    expect(
+      walletFinalBalance.nRLC
+        .add(new BN('5000000'))
+        .eq(walletInitialBalance.nRLC),
     ).toBe(true);
     expect(accountFinalBalance.locked.eq(accountInitialBalance.locked)).toBe(
       true,
@@ -1806,6 +1984,38 @@ describe('[account]', () => {
       true,
     );
   });
+  test('account.withdraw() (token, specified unit)', async () => {
+    const signer = utils.getSignerFromPrivateKey(tokenChainUrl, PRIVATE_KEY);
+    const iexec = new IExec(
+      {
+        ethProvider: signer,
+        chainId: networkId,
+      },
+      {
+        hubAddress,
+        isNative: false,
+      },
+    );
+    await iexec.account.deposit(10000);
+    const accountInitialBalance = await iexec.account.checkBalance(ADDRESS);
+    const walletInitialBalance = await iexec.wallet.checkBalances(ADDRESS);
+    const res = await iexec.account.withdraw('0.000005 RLC');
+    const accountFinalBalance = await iexec.account.checkBalance(ADDRESS);
+    const walletFinalBalance = await iexec.wallet.checkBalances(ADDRESS);
+    expect(res.txHash).toMatch(bytes32Regex);
+    expect(res.amount).toBe('5000');
+    expect(
+      accountFinalBalance.stake
+        .add(new BN(5000))
+        .eq(accountInitialBalance.stake),
+    ).toBe(true);
+    expect(
+      walletFinalBalance.nRLC.sub(new BN(5000)).eq(walletInitialBalance.nRLC),
+    ).toBe(true);
+    expect(accountFinalBalance.locked.eq(accountInitialBalance.locked)).toBe(
+      true,
+    );
+  });
   test('account.withdraw() (token, exceed account balance)', async () => {
     const signer = utils.getSignerFromPrivateKey(tokenChainUrl, PRIVATE_KEY);
     const iexec = new IExec(
@@ -1860,6 +2070,38 @@ describe('[account]', () => {
     ).toBe(true);
     expect(
       walletFinalBalance.nRLC.sub(new BN(5)).eq(walletInitialBalance.nRLC),
+    ).toBe(true);
+    expect(accountFinalBalance.locked.eq(accountInitialBalance.locked)).toBe(
+      true,
+    );
+  });
+  test('account.withdraw() (native, specified unit)', async () => {
+    const signer = utils.getSignerFromPrivateKey(nativeChainUrl, PRIVATE_KEY);
+    const iexec = new IExec(
+      {
+        ethProvider: signer,
+        chainId: networkId,
+      },
+      {
+        hubAddress: nativeHubAddress,
+        isNative: true,
+      },
+    );
+    await iexec.account.deposit(10000);
+    const accountInitialBalance = await iexec.account.checkBalance(ADDRESS);
+    const walletInitialBalance = await iexec.wallet.checkBalances(ADDRESS);
+    const res = await iexec.account.withdraw('0.000005 RLC');
+    const accountFinalBalance = await iexec.account.checkBalance(ADDRESS);
+    const walletFinalBalance = await iexec.wallet.checkBalances(ADDRESS);
+    expect(res.txHash).toMatch(bytes32Regex);
+    expect(res.amount).toBe('5000');
+    expect(
+      accountFinalBalance.stake
+        .add(new BN(5000))
+        .eq(accountInitialBalance.stake),
+    ).toBe(true);
+    expect(
+      walletFinalBalance.nRLC.sub(new BN(5000)).eq(walletInitialBalance.nRLC),
     ).toBe(true);
     expect(accountFinalBalance.locked.eq(accountInitialBalance.locked)).toBe(
       true,
@@ -2472,17 +2714,54 @@ describe('[order]', () => {
         isNative: false,
       },
     );
+    const app = getRandomAddress();
     const order = await iexec.order.createApporder({
-      app: POOR_ADDRESS2,
+      app,
     });
     expect(order).toEqual({
-      app: '0x650ae1d365369129c326Cd15Bf91793b52B7cf59',
+      app,
       appprice: '0',
       datasetrestrict: '0x0000000000000000000000000000000000000000',
       requesterrestrict: '0x0000000000000000000000000000000000000000',
       tag: '0x0000000000000000000000000000000000000000000000000000000000000000',
       volume: '1',
       workerpoolrestrict: '0x0000000000000000000000000000000000000000',
+    });
+  });
+
+  test('order.createApporder() (override defaults)', async () => {
+    const signer = utils.getSignerFromPrivateKey(tokenChainUrl, PRIVATE_KEY);
+    const iexec = new IExec(
+      {
+        ethProvider: signer,
+        chainId: networkId,
+      },
+      {
+        hubAddress,
+        isNative: false,
+      },
+    );
+    const app = getRandomAddress();
+    const datasetrestrict = getRandomAddress();
+    const workerpoolrestrict = getRandomAddress();
+    const requesterrestrict = getRandomAddress();
+    const order = await iexec.order.createApporder({
+      app,
+      appprice: '1 RLC',
+      datasetrestrict,
+      workerpoolrestrict,
+      requesterrestrict,
+      tag: 'tee',
+      volume: 100,
+    });
+    expect(order).toEqual({
+      app,
+      appprice: '1000000000',
+      datasetrestrict,
+      requesterrestrict,
+      tag: '0x0000000000000000000000000000000000000000000000000000000000000001',
+      volume: '100',
+      workerpoolrestrict,
     });
   });
 
@@ -2498,17 +2777,54 @@ describe('[order]', () => {
         isNative: false,
       },
     );
+    const dataset = getRandomAddress();
     const order = await iexec.order.createDatasetorder({
-      dataset: POOR_ADDRESS2,
+      dataset,
     });
     expect(order).toEqual({
       apprestrict: '0x0000000000000000000000000000000000000000',
-      dataset: '0x650ae1d365369129c326Cd15Bf91793b52B7cf59',
+      dataset,
       datasetprice: '0',
       requesterrestrict: '0x0000000000000000000000000000000000000000',
       tag: '0x0000000000000000000000000000000000000000000000000000000000000000',
       volume: '1',
       workerpoolrestrict: '0x0000000000000000000000000000000000000000',
+    });
+  });
+
+  test('order.createDatasetorder() (override defaults)', async () => {
+    const signer = utils.getSignerFromPrivateKey(tokenChainUrl, PRIVATE_KEY);
+    const iexec = new IExec(
+      {
+        ethProvider: signer,
+        chainId: networkId,
+      },
+      {
+        hubAddress,
+        isNative: false,
+      },
+    );
+    const dataset = getRandomAddress();
+    const apprestrict = getRandomAddress();
+    const workerpoolrestrict = getRandomAddress();
+    const requesterrestrict = getRandomAddress();
+    const order = await iexec.order.createDatasetorder({
+      dataset,
+      datasetprice: '1 RLC',
+      apprestrict,
+      workerpoolrestrict,
+      requesterrestrict,
+      tag: 'tee',
+      volume: 100,
+    });
+    expect(order).toEqual({
+      dataset,
+      datasetprice: '1000000000',
+      apprestrict,
+      requesterrestrict,
+      tag: '0x0000000000000000000000000000000000000000000000000000000000000001',
+      volume: '100',
+      workerpoolrestrict,
     });
   });
 
@@ -2524,8 +2840,9 @@ describe('[order]', () => {
         isNative: false,
       },
     );
+    const workerpool = getRandomAddress();
     const order = await iexec.order.createWorkerpoolorder({
-      workerpool: POOR_ADDRESS2,
+      workerpool,
       category: 5,
     });
     expect(order).toEqual({
@@ -2536,8 +2853,48 @@ describe('[order]', () => {
       tag: '0x0000000000000000000000000000000000000000000000000000000000000000',
       trust: '0',
       volume: '1',
-      workerpool: '0x650ae1d365369129c326Cd15Bf91793b52B7cf59',
+      workerpool,
       workerpoolprice: '0',
+    });
+  });
+
+  test('order.createWorkerpoolorder() (override defaults)', async () => {
+    const signer = utils.getSignerFromPrivateKey(tokenChainUrl, PRIVATE_KEY);
+    const iexec = new IExec(
+      {
+        ethProvider: signer,
+        chainId: networkId,
+      },
+      {
+        hubAddress,
+        isNative: false,
+      },
+    );
+    const workerpool = getRandomAddress();
+    const apprestrict = getRandomAddress();
+    const datasetrestrict = getRandomAddress();
+    const requesterrestrict = getRandomAddress();
+    const order = await iexec.order.createWorkerpoolorder({
+      workerpool,
+      workerpoolprice: '0.1 RLC',
+      category: 5,
+      apprestrict,
+      datasetrestrict,
+      requesterrestrict,
+      tag: 'tee',
+      trust: '10',
+      volume: '100',
+    });
+    expect(order).toEqual({
+      apprestrict,
+      category: '5',
+      datasetrestrict,
+      requesterrestrict,
+      tag: '0x0000000000000000000000000000000000000000000000000000000000000001',
+      trust: '10',
+      volume: '100',
+      workerpool,
+      workerpoolprice: '100000000',
     });
   });
 
@@ -2554,12 +2911,13 @@ describe('[order]', () => {
         resultProxyURL: 'https://result-proxy.iex.ec',
       },
     );
+    const app = getRandomAddress();
     const order = await iexec.order.createRequestorder({
-      app: POOR_ADDRESS2,
+      app,
       category: 5,
     });
     expect(order).toEqual({
-      app: '0x650ae1d365369129c326Cd15Bf91793b52B7cf59',
+      app,
       appmaxprice: '0',
       beneficiary: ADDRESS,
       callback: '0x0000000000000000000000000000000000000000',
@@ -2578,6 +2936,65 @@ describe('[order]', () => {
       volume: '1',
       workerpool: '0x0000000000000000000000000000000000000000',
       workerpoolmaxprice: '0',
+    });
+  });
+
+  test('order.createRequestorder() (override defaults)', async () => {
+    const signer = utils.getSignerFromPrivateKey(tokenChainUrl, PRIVATE_KEY);
+    const iexec = new IExec(
+      {
+        ethProvider: signer,
+        chainId: networkId,
+      },
+      {
+        hubAddress,
+        isNative: false,
+        resultProxyURL: 'https://result-proxy.iex.ec',
+      },
+    );
+    const app = getRandomAddress();
+    const dataset = getRandomAddress();
+    const workerpool = getRandomAddress();
+    const callback = getRandomAddress();
+    const order = await iexec.order.createRequestorder({
+      app,
+      category: 5,
+      dataset,
+      workerpool,
+      callback,
+      appmaxprice: '1 nRLC',
+      datasetmaxprice: '100 nRLC',
+      workerpoolmaxprice: '0.1 RLC',
+      params: {
+        iexec_result_storage_provider: 'dropbox',
+        iexec_tee_post_compute_fingerprint: 'teePostComputeFingerprint',
+        iexec_tee_post_compute_image: 'teePostComputeImage',
+        iexec_result_encryption: true,
+      },
+      tag: 'tee',
+      trust: '100',
+      volume: '5',
+    });
+    expect(order).toEqual({
+      app,
+      appmaxprice: '1',
+      beneficiary: ADDRESS,
+      callback,
+      category: '5',
+      dataset,
+      datasetmaxprice: '100',
+      params: {
+        iexec_result_storage_provider: 'dropbox',
+        iexec_tee_post_compute_fingerprint: 'teePostComputeFingerprint',
+        iexec_tee_post_compute_image: 'teePostComputeImage',
+        iexec_result_encryption: true,
+      },
+      requester: ADDRESS,
+      tag: '0x0000000000000000000000000000000000000000000000000000000000000001',
+      trust: '100',
+      volume: '5',
+      workerpool,
+      workerpoolmaxprice: '100000000',
     });
   });
 
@@ -6115,7 +6532,7 @@ describe('[lib utils]', () => {
       const encZip = await fs.readFile(
         path.join(
           process.cwd(),
-          'test/inputs/encryptedResults/encryptedTeeRes.zip',
+          'test/inputs/encryptedResults/encryptedResults.zip',
         ),
       );
       const beneficiaryKey = await fs.readFile(
@@ -6130,15 +6547,16 @@ describe('[lib utils]', () => {
       resZip.forEach((relativePath, zipEntry) => {
         resContent.push(zipEntry);
       });
-      expect(resContent.length).toBe(2);
-      expect(resContent[0].name).toBe('volume.fspf');
-      expect(resContent[1].name).toBe('result.png');
+      expect(resContent.length).toBe(3);
+      expect(resContent[0].name).toBe('computed.json');
+      expect(resContent[1].name).toBe('volume.fspf');
+      expect(resContent[2].name).toBe('result.txt');
     });
     test('result.decryptResult() string key', async () => {
       const encZip = await fs.readFile(
         path.join(
           process.cwd(),
-          'test/inputs/encryptedResults/encryptedTeeRes.zip',
+          'test/inputs/encryptedResults/encryptedResults.zip',
         ),
       );
       const beneficiaryKey = (
@@ -6155,15 +6573,16 @@ describe('[lib utils]', () => {
       resZip.forEach((relativePath, zipEntry) => {
         resContent.push(zipEntry);
       });
-      expect(resContent.length).toBe(2);
-      expect(resContent[0].name).toBe('volume.fspf');
-      expect(resContent[1].name).toBe('result.png');
+      expect(resContent.length).toBe(3);
+      expect(resContent[0].name).toBe('computed.json');
+      expect(resContent[1].name).toBe('volume.fspf');
+      expect(resContent[2].name).toBe('result.txt');
     });
     test('result.decryptResult() wrong key', async () => {
       const encZip = await fs.readFile(
         path.join(
           process.cwd(),
-          'test/inputs/encryptedResults/encryptedTeeRes.zip',
+          'test/inputs/encryptedResults/encryptedResults.zip',
         ),
       );
       const beneficiaryKey = await fs.readFile(