+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
+## [5.0.1] 2020-08-26
+
+### Added
+
+- support for unit in amounts
+  - ether units: `ether` (`eth`), `finney`, `szabo`, `gwei`, `mwei`, `kwei`, `wei`
+  - RLC units: `RLC`, `nRLC`
+
+### Changed
+
+- fixed result encryption v5 workflow
+- messages amount use main units (RLC and ether)
+- [DEPRECATED] `iexec wallet show --raw` returned json key `balance.ETH` is deprecated, use `balance.ether` instead
+
+### Removed
+
 ## [5.0.0] 2020-07-22
 
 ### Added